pyproject.toml
[tool.poetry]
name = "pi-light"
version = "0.1.0"
description = "Django WebApp for command Blinkt Light on Raspberry Pi"
authors = ["obviousrebel <joshalex359@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.5"
django-material = "^1.6.3"
django-viewflow = "^1.6.0"
django-colorful = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
